+++ react/interfaces/custom-pixel-event.interface.ts
@@ -1,0 +1,11 @@
+export interface CustomPixelEventInterface {
+  [key: string]: CustomPixelEventPropTypes | CustomPixelEventPropTypes[]
+}
+
+type CustomPixelEventPropTypes =
+  | string
+  | number
+  | boolean
+  | Record<string, unknown>
+  | unknown[]
+  | null